+++ lib/main.dart
@@ -28,7 +28,7 @@ void initAsync(MainAppBloc bloc) async {
   Log.debug('this is initAsync');
   // await Future.delayed(const Duration(seconds: 5));
   final Mode mode = kReleaseMode ? Mode.Production : Mode.Debug;
-  final List<Module> modules = kReleaseMode ? <Module>[ProdModule()] : <Module>[DevModule()];
+  final List<Module> modules = kReleaseMode ? <Module>[ProdModule()] : <Module>[DevModuleCore()];
 
   Config.initAsync(mode)
       .then((_) => DI.initAsync(modules))
