+++ petisland_core/lib/module/prod_module_core.dart
@@ -1,0 +1,3 @@
+part of petisland_core.module;
+
+class ProductionModuleCore extends DevModuleCore {}
