+++ petisland_core/lib/petisland_core.dart
@@ -3,4 +3,5 @@ library petisland_core;
 export 'client/client.dart';
 export 'config/config.dart';
 export 'exception/exception.dart';
+export 'module/module.dart';
 export 'util/log.dart';
